+++ incubating/argo-cd-sync/CHANGELOG.md
@@ -1,7 +1,18 @@
 # Changelog
-## [1.3.1] - 2023-09-18
-### Changed
 
+## [1.4.2] - 2024-01-17
+### Changed
+New graphql call to speed up query
+
+## [1.4.1] - 2023-10-31
+### Changed
+Add CA_BUNDLE option
+
+## [1.4.0] - 2023-10-30
+### Changed
+Add INSECURE option
+
+## [1.3.1] - 2023-09-18
 ### Fixed
 - CVE-2023-37920 - upgrade Python module certifi to 2023.7.22
 - CVE-2019-8457 - upgrade base image to python:3.11.5-slim-bookworm
@@ -10,5 +21,3 @@
 ### Changed
 - Adding IMAGE_NAME parameter
 - Adding example
-
-### Fixed